embedding/requirements.txt
langchain
openai==0.28.1
python-dotenv
chromadb
tiktoken